Move quiz data out to its own file
